locale/zh-CN: apply common translation for 'open with' for consistency
In windows and various FOSS platforms, 'open with' is translated as 打开方式,
literally 'method to open'. This commit tries to apply this convention, to
menu items and to the program name.

Translating the program name is done for consistency. Normally I don't do that,
but for this particular program it sounds better to do so, given that the
phrase is so common...

To avoid ambiguity places where "Open With" appear as an extension name have
quotes around the translated string, i.e. "“打开方式”". See #97 (PR) for more
discussion on this decision.