Added ungoogled-chromium installation support.
Previously, the native messaging host manifest file wouldn't install to
ungoogled-chromium's NativeMessagingHosts directory, so to make
open-with work, one had to manually copy the manifest file into this
directory. With this commit, the installer now supports installing the
manifest file on ungoogled-chromium. Note that this only works on Linux.
This commits also adds detection for ungoogled-chromium when
looking for browsers.